extract titlepage content page
