+++ modules/content/material/chunks/veteran/chunk.md
@@ -2,6 +2,8 @@
 id: "veteran"
 ---
 
-Welcome back! Hey, since you know your way around, please {EXTLINK:url=mailtoCOLONdavid@krawaller.se,text=mail us,style=link} or {DYNAMIC:id=slack,style=link} if you have any feedback.
+Welcome back! Our latest game is {DYNAMIC:id=latest,style=link}, have you played that yet?
+
+Hey, since you know your way around, please {EXTLINK:url=mailtoCOLONdavid@krawaller.se,text=mail us,style=link} or {DYNAMIC:id=slack,style=link,text=join the Slack} if you have any feedback.
 
 Or, if we're <em>very</em> good friends, maybe even {EXTLINK:text=buy us a coffee,url=https://www.buymeacoffee.com/chessicals,style=link} to help further development! 😊
